path refactor: move shared plugin code to sessions and adobe_shared libs; add missing CMakeLists; branch now builds
