+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.12.0 FATAL_ERROR)
+cmake_minimum_required(VERSION 3.15.0 FATAL_ERROR)
 
 # modules
 LIST(APPEND CMAKE_MODULE_PATH "${CMAKE_SOURCE_DIR}/external/cmake-pandocology")
@@ -23,5 +23,4 @@ add_subdirectory(external)
 # add_subdirectory(doc)
 
 # plugins
-add_subdirectory(source/premiere_CC2018)
 add_subdirectory(source)
