+++ source/plugin/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.15.0 FATAL_ERROR)
+
+# plugins
+add_subdirectory(adobe_shared)
+add_subdirectory(after_effects)
+add_subdirectory(premiere)
